backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.3.0
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
librosa==0.10.1
numpy==1.25.2
pandas==2.1.0
scikit-learn==1.3.0
torch==2.0.1
torchvision==0.15.2
transformers==4.32.1
pillow==10.0.0
sentence-transformers==2.2.2
python-dotenv==1.0.0